requirements.txt
Appium-Python-Client==3.2.1
attrs==23.2.0
browserstack-local==1.2.7
browserstack-sdk==1.19.24
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
docutils==0.20.1
dukpy==0.3.1
exceptiongroup==1.2.0
execnet==2.1.1
filelock==3.13.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.6
iniconfig==2.0.0
kitchen==1.2.6
natsort==8.4.0
outcome==1.3.0.post0
packaging==24.0
percy-appium-app==2.0.5
pluggy==1.4.0
psutil==5.9.8
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-SystemConfiguration==10.2
PyPAC==0.16.4
PySocks==1.7.1
pytest==8.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
robotframework==7.0
robotframework-appiumlibrary==2.1.0
robotframework-pabot==2.18.0
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.2.0
robotframework-stacktrace==0.4.1
robotframework-requests==0.9.4
selenium==4.16.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.2
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
Faker==8.0.0